Weird randomPoints.h redefinition error
